CheckCollision() is busted and I can't figure it out
Alright so here's the deal. I've been looking at the return value of CheckCollision, wondering why it isn't just returning the map data like it should.

When I printf right in the function, it shows the correct map info. But when returned, it is all wrong and I do not know why!